Refactored TeamCreate and added TeamBuilder component to display the builder aspect
